refactor: move VCARD processing to process_message and add data_access __init__ modules
