+++ backend/app/db/data_access/user_goals/__init__.py
@@ -1,0 +1,8 @@
+"""User goals data access module.
+
+This module exports functions for retrieving user goals from the database.
+"""
+from .get_user_goals import get_user_goals
+
+__all__ = ["get_user_goals"]
+
