+++ backend/app/logic/helpers/format_week_summary.py
@@ -5,7 +5,7 @@ goal tracking summaries for users.
 """
 from datetime import datetime, timedelta
 from app.logic.helpers.look_back_summary import look_back_summary
-from app.db.data_access.user_goals.get_user_goals import get_user_goals
+from app.db.data_access import get_user_goals
 
 def get_monday_before(date: datetime | None = None) -> datetime:
     """
