+++ backend/app/logic/helpers/look_back_summary.py
@@ -6,7 +6,7 @@ of goal ratings over specified time periods.
 from datetime import datetime, timedelta
 from app.db.sqlite import get_db
 from app.utils.config import STYLE
-from app.db.data_access.user_goals.get_user_goals import get_user_goals
+from app.db.data_access import get_user_goals
 from app.utils.messages import (
     LOOKBACK_NO_GOALS,
     LOOKBACK_USER_NOT_FOUND,
