+++ backend/app/logic/helpers/rate_individual_goal.py
@@ -6,7 +6,7 @@ with validation and database storage.
 from datetime import datetime
 from app.db.sqlite import get_db
 from app.utils.config import STYLE
-from app.db.data_access.user_goals.get_user_goals import get_user_goals
+from app.db.data_access import get_user_goals
 from app.utils.messages import (
     ERROR_RATING_INVALID,
     ERROR_NO_GOALS_ADD_FIRST,
